Add CLI entrypoints for MCP Docs and Systems servers
- Implemented `main.py` for the Deephaven MCP Docs server, providing a command-line interface to start the server with specified transports (stdio, sse, streamable-http).
- Created a new `main.py` for the Deephaven MCP Systems server, similar to the Docs server, allowing for transport specification.
- Removed the `run_server` function from the `mcp_systems_server` package's `__init__.py` and updated the documentation accordingly.
- Updated tests for the MCP Docs and Systems servers to cover the new CLI entrypoints and ensure proper logging and exception handling.
- Refactored logging setup to ensure it is configured before server execution.
- Added tests for logging setup and exception handling in the logging module.
- Cleaned up unused tests and imports in the MCP Docs and Systems server test files.
